Array Rotation Returns (Problem Code: ARRT)
In this, I have provided an elegant solution to the problem that appeared in July Lunchtime 2021 Div-3 (Problem - 4). Here's the link to the problem - https://www.codechef.com/LTIME98C/problems/ARRT/
I would encourage you to try this on your own for the first time and then go through the solution.